+++ src/main/java/lotto/domain/lotto/LottoNumber.java
@@ -1,0 +1,47 @@
+package lotto.domain.lotto;
+
+import java.util.Objects;
+
+class LottoNumber implements Comparable<LottoNumber> {
+
+    static final int MINIMUM_LOTTO_NUMBER = 1;
+    static final int MAXIMUM_LOTTO_NUMBER = 45;
+
+    private final int number;
+
+    LottoNumber(int number) {
+        validateLottoNumber(number);
+        this.number = number;
+    }
+
+    private void validateLottoNumber(int number) {
+        if (number < MINIMUM_LOTTO_NUMBER || number > MAXIMUM_LOTTO_NUMBER) {
+            throw new IllegalArgumentException("로또는 1부터 45까지의 숫자만 가능합니다.");
+        }
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)
+            return true;
+        if (o == null || getClass() != o.getClass())
+            return false;
+        LottoNumber that = (LottoNumber) o;
+        return number == that.number;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(number);
+    }
+
+    @Override
+    public String toString() {
+        return String.valueOf(number);
+    }
+
+    @Override
+    public int compareTo(LottoNumber targetNumber) {
+        return this.number - targetNumber.number;
+    }
+}
